Move `loadAudio` to utils file
