+++ tests/pipelines.test.js
@@ -1,7 +1,7 @@
 
 import { pipeline, cos_sim } from '../src/transformers.js';
 import { init, m, MAX_TEST_EXECUTION_TIME } from './init.js';
-import { compare } from './test_utils.js';
+import { compare, loadAudio } from './test_utils.js';
 
 // Initialise the testing environment
 init();
@@ -704,26 +704,6 @@ describe('Pipelines', () => {
 
     describe('Speech-to-text generation', () => {
 
-        const loadAudio = async (url) => {
-            // NOTE: Since the Web Audio API is not available in Node.js, we will need to use the `wavefile` library to obtain the raw audio data.
-            // For more information, see: https://huggingface.co/docs/transformers.js/tutorials/node-audio-processing
-            let wavefile = (await import('wavefile')).default;
-
-            // Load audio data
-            let buffer = Buffer.from(await fetch(url).then(x => x.arrayBuffer()))
-
-            // Read .wav file and convert it to required format
-            let wav = new wavefile.WaveFile(buffer);
-            wav.toBitDepth('32f'); // Pipeline expects input as a Float32Array
-            wav.toSampleRate(16000); // Whisper expects audio with a sampling rate of 16000
-            let audioData = wav.getSamples();
-            if (Array.isArray(audioData)) {
-                // For this demo, if there are multiple channels for the audio file, we just select the first one.
-                // In practice, you'd probably want to convert all channels to a single channel (e.g., stereo -> mono).
-                audioData = audioData[0];
-            }
-            return audioData;
-        }
         // List all models which will be tested
         const models = [
             'openai/whisper-tiny.en', // English-only
